+++ movement/napari/meta_widget.py
@@ -3,7 +3,7 @@
 from napari.viewer import Viewer
 from qt_niu.collapsible_widget import CollapsibleWidgetContainer
 
-from movement.napari.widgets import PosesLoader
+from movement.napari.loader_widgets import PosesLoader
 
 
 class MovementMetaWidget(CollapsibleWidgetContainer):
