+++ movement/napari/napari.yaml
@@ -3,7 +3,7 @@ display_name: movement
 contributions:
   commands:
     - id: movement.make_widget
-      python_name: movement.napari._meta_widget:MovementMetaWidget
+      python_name: movement.napari.meta_widget:MovementMetaWidget
       title: movement
   widgets:
     - command: movement.make_widget
